moved api file to title-route.js and created a route to display title json data
